start refactoring pydantic models for each object into their own folders
